feat(git): add plugin config for gitsigns and disable open git diff mapping
Issue here is that the git signs plugin is initialised on the User
AstroGitFile event which I'm certain is triggered after loading the
custom mappings, so my mapping for diffview is overriden. I could
declare the diffview plugin after this with the same event and add my
mappings there, that should resolve this.